uiuuliuli

+++ Controllers/Order_details.php
@@ -1,16 +1,20 @@
 <?php
 
 //delete
-if (isset($_POST['del_id'])) {
+if (!isset($_SESSION['admin'])) {
+    header("Location:./index.php" );
+}else{
+    if (isset($_POST['del_id'])) {
 
-    $del_id = $_POST['del_id'];
-    
-    if ($OrderDetails->deleteOrder($del_id)) {
-       echo "deleted";
-        // RELOAD PAGE
-        header("Location:" . $_SERVER['PHP_SELF']);
-    }else {
-        die('failed');
+        $del_id = $_POST['del_id'];
+        
+        if ($OrderDetails->deleteOrder($del_id)) {
+        echo "deleted";
+            // RELOAD PAGE
+            header("Location:" . $_SERVER['PHP_SELF']);
+        }else {
+            die('failed');
+        }
     }
 }
